refactor(services): shrink cache fetcher function and split job model module
